requirements.txt
streamlit
pandas
numpy
psycopg2-binary
pdfkit
altair
matplotlib
python-dotenv
plotly
openpyxl